added: transfer near to toList
